Q1_HR_Assistant/requirements.txt
streamlit
chromadb
openai>=1.0.0
python-dotenv
pdfplumber
python-docx
sentence-transformers
